Move formatSaveError to production module
- Extract error formatting utility from test file to lib/kb/error-handling.ts
- Update test imports to use the production module
- Fix service worker consistency: use cloned request for cache.put()

Addresses feedback from Cursor Bugbot and Claude review.
